WIP implementation of arch components lifecycle support
This is a first pass of support for Android Architecture Components' `Lifecycle` support. This works fairly similarly to `ViewScopeProvider` from a structural standpoint. `LifecycleOwner` or `Lifecycle` can be passed in, an inner observer is passed in and removed upon disposal.

Thoughts for discussion:
  - Name is up for grabs. This one's a bit long :)
  - Should it be its own artifact? I'm leaning toward this, then we could also do an alpha release of it separately.
  - Things get a bit convoluted for me around `State` vs. `Event`. I'm not sure which one we should watch for, but all the querying APIs revolve around `State` so I went with that.
  - Does it matter if it's on the main thread?

If this looks good - I'll take a pass at tests.